HLFIR: route scalar LOGICAL inputs as np.bool_ (silences int32->bool* warning)
The cloudsc registry's six scalar LOGICAL parameters (LAERLIQCOLL,
LDMAINCALL, LDSLPHY, LAERLIQAUTOCP, LAERLIQAUTOCPB, LAERLIQAUTOLSP,
LAERICESED, LAERICEAUTO, LCLDEXTRA, LCLDBUDGET) were stored as
np.int32(np.bool_(True)) -> 4-byte int.  The bridge declares them as
bool * (1-byte) on the C ABI, so the DaCe runtime warned:
   'Passing int32 array argument X to a bool array'
and reinterpret-cast the buffer.  For value=1 the int32 LSB happens to
read as true in C bool, but ANY value with bit-0 = 0 (e.g. 256, -2)
would silently read as false.

Fixes:
- Registry now stores scalar LOGICALs as np.bool_(...) directly.
- test_cloudsc_full.py::_sdfg_call_args dispatches by the bridge's
  declared dtype (bool / float / int) so callers can pass any
  compatible Python value without picking the dtype themselves.
- Adds test_bool_scalar_logical_pass_through to bool_int_arg_dtype_test.py
  -- pins the scalar-LOGICAL contract that the prior 1D/2D array
  tests missed.

No numerical impact on cloudsc_full (the 1351 mismatching cells stay
the same -- LOGICAL values were all True in the test, where the int32
LSB coincidentally reads correctly).  Removes the three runtime
warnings.

+++ tests/hlfir/cloudsc_full/_registries.py
@@ -50,16 +50,21 @@ parameters = {
     'NCLDDIAG': 1,
     'NAERCLD': 1,
     'NBLOCKS': nbvalue,
-    'LDMAINCALL': np.int32(np.bool_(True)),
-    'LDSLPHY': np.int32(np.bool_(True)),
-    'LAERLIQAUTOCP': np.int32(np.bool_(True)),
-    'LAERLIQAUTOCPB': np.int32(np.bool_(True)),
-    'LAERLIQAUTOLSP': np.int32(np.bool_(True)),
-    'LAERLIQCOLL': np.int32(np.bool_(True)),
-    'LAERICESED': np.int32(np.bool_(True)),
-    'LAERICEAUTO': np.int32(np.bool_(True)),
-    'LCLDEXTRA': np.int32(np.bool_(True)),
-    'LCLDBUDGET': np.int32(np.bool_(True)),
+    # Scalar LOGICAL parameters: keep as np.bool_ so the call-site
+    # routing in test_cloudsc_full.py::_sdfg_call_args picks the
+    # bridge's ``bool *`` ABI cleanly.  Casting to np.int32 here would
+    # work for value=1 by accident (LSB matches) but silently corrupt
+    # any value with bit-0 = 0 -- see ``test_bool_scalar_logical_pass_through``.
+    'LDMAINCALL': np.bool_(True),
+    'LDSLPHY': np.bool_(True),
+    'LAERLIQAUTOCP': np.bool_(True),
+    'LAERLIQAUTOCPB': np.bool_(True),
+    'LAERLIQAUTOLSP': np.bool_(True),
+    'LAERLIQCOLL': np.bool_(True),
+    'LAERICESED': np.bool_(True),
+    'LAERICEAUTO': np.bool_(True),
+    'LCLDEXTRA': np.bool_(True),
+    'LCLDBUDGET': np.bool_(True),
     'NGPBLKS': 10,
     'NUMOMP': 10,
     'NGPTOT': nbvalue * 1,
